Projeto Final Finalizado

+++ dump.sql
@@ -46,7 +46,7 @@ CREATE TABLE
         id SERIAL PRIMARY KEY,
         cliente_id INT REFERENCES clientes(id),
         observacao TEXT,
-        valor_total INT
+        valor_total FLOAT
     );
 
 CREATE TABLE
